Cleanup of Code & File Structure
