+++ RPG Wizardry/Assets/Art/Animations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 345786e64ca8a7f4895c50aa5f25573a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
