+++ RPG Wizardry/Assets/Art/Animations/Entities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6e3ac24076ee35b4898e144b62f181a7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
