+++ RPG Wizardry/Assets/Art/Animations/Entities/Avatar.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 589ded06bc3d4d84c951b3e0d956059e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
