+++ RPG Wizardry/Assets/Art/Animations/Entities/Avatar/willowEast.anim.meta
