+++ RPG Wizardry/Assets/Art/Animations/Entities/Avatar/willowEastIdle.anim
