+++ RPG Wizardry/Assets/Art/Animations/Entities/Avatar/willowNorth.anim
