+++ RPG Wizardry/Assets/Art/Animations/Entities/Avatar/willowNorthIdle.anim.meta
