+++ RPG Wizardry/Assets/Art/Animations/Entities/Avatar/willowSouth.anim.meta
