+++ RPG Wizardry/Assets/Art/Animations/Entities/Avatar/willowSouthIdle.anim.meta
