+++ RPG Wizardry/Assets/Art/Animations/Entities/Avatar/willowWest.anim
