+++ RPG Wizardry/Assets/Art/Animations/Entities/Avatar/willowWestIdle.anim
