+++ RPG Wizardry/Assets/Art/Sprites/Entities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 84eb5d97166d9a743a5bd909be2059c1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
