+++ RPG Wizardry/Assets/Art/Sprites/Entities/Avatar.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4f61738649a6ba7429d766d1ec053985
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
