+++ RPG Wizardry/Assets/Art/Sprites/Environment.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8d9a369b95394294cac78c863864f5e6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
