--- RPG Wizardry/Assets/Scripts/ArrayExtensions.cs
@@ -1,33 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Runtime.InteropServices;
-using UnityEngine;
-
-public static class ArrayExtensions
-{
-    public static T[] GetRow<T>(this T[,] array, int row)
-    {
-        if (!typeof(T).IsPrimitive)
-            throw new InvalidOperationException("Not supported for managed types.");
-
-        if (array == null)
-            throw new ArgumentNullException("array");
-
-        int cols = array.GetUpperBound(1) + 1;
-        T[] result = new T[cols];
-
-        int size;
-
-        if (typeof(T) == typeof(bool))
-            size = 1;
-        else if (typeof(T) == typeof(char))
-            size = 2;
-        else
-            size = Marshal.SizeOf<T>();
-
-        Buffer.BlockCopy(array, row * cols * size, result, 0, cols * size);
-
-        return result;
-    }
-}
