+++ RPG Wizardry/Assets/Scripts/Avatar.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d22ba39dc8bf93843a4fdaf650a5c877
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
