+++ RPG Wizardry/Assets/Scripts/Avatar/Movement.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 59bdf4a6170dc8e43a862540e583d0a1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
