+++ RPG Wizardry/Assets/Scripts/GameLogic.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9c72b2a9ec090034cb0f6f4aa10e022d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
