+++ RPG Wizardry/Assets/Scripts/GameLogic/GameState.cs
@@ -1,0 +1,8 @@
+﻿namespace nl.SWEG.RPGWizardry.GameLogic
+{
+    public enum GameState
+    {
+        Menu = 0,
+        GamePlay = 1
+    }
+}
