+++ RPG Wizardry/Assets/Scripts/GameLogic/GameState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9cb5343125382244c9cf582470f927c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
