+++ RPG Wizardry/Assets/Scripts/PlayerInput.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9746574fa7fd08a47bbd88749283b27a
+guid: 0bddc6714413d89439efe48cddd8a27c
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
