+++ RPG Wizardry/Assets/Scripts/PlayerInput/SpellCraftingInput.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3b7487891f148454d9e4f73507b90db1
+guid: b0d9aeda537dcb44d8e0d0fb3cbb0ac2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
