+++ RPG Wizardry/Assets/Scripts/SpellCrafting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 98e8847ef58125146b035262a309f883
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
