+++ RPG Wizardry/Assets/Scripts/Utils/Direction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9ef8774f1f750e44aa37105fa4b81760
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
